Move fleet to fleet/grpc_rere (#1862)
